pyproject.toml
[project]
name = "sbir-search"
version = "0.1.0"
description = "Nightly SBIR opportunity crawler focused on AI-powered reverse engineering tools"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "blacktop" }]
keywords = ["sbir", "sttr", "grants", "opportunities", "crawler", "discord"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Office/Business",
]

dependencies = [
  "discord.py>=2.0",
  "feedparser>=6.0",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.8",
]

[project.scripts]
sbir-search = "sbir_search.cli:main"

[project.urls]
Repository = "https://github.com/blacktop/sbir-search"

[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sbir_search"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "C4", "PIE", "PT", "RUF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
